Make zsh autocomplete case sensitive so that ENV vars don't keep popping up
There doesn't seem to be a way to ignore env vars from autocomplete, and
getting a screen full of env var possibilities when trying to autocomplete
something like zsh-* is a pain. Hence, turning case sensitive autocomplete
on seems to be the lesser of two evils. Refer: https://github.com/ohmyzsh/ohmyzsh/issues/3885.